+++ packages/web-components/src/components/style-picker/__stories__/style-picker.flat.stories.js
@@ -21,6 +21,9 @@ import '@carbon/web-components/es/components/tile/index.js';
 import '@carbon/web-components/es/components/link/index.js';
 import { BUTTON_KIND } from '@carbon/web-components/es/components/button/defs.js';
 
+// Pictograms
+import AmsterdamWindmill from '@carbon/pictograms/lib/amsterdam--windmill';
+
 // Carbon labs
 import { settings } from '@carbon-labs/utilities/es/settings/index.js';
 
@@ -76,9 +79,18 @@ const styles = css`
  * remove and add `open` attribute
  */
 const toggleButton = () => {
-  document
-    .querySelector(`${clabsPrefix}-style-picker`)
-    ?.toggleAttribute('open');
+  const stylePicker = document.querySelector(`${clabsPrefix}-style-picker`);
+  const trigger = document.querySelector('#trigger');
+
+  stylePicker?.toggleAttribute('open');
+  trigger.setAttribute('aria-expanded', !!stylePicker?.hasAttribute('open'));
+};
+
+/**
+ * Event fire when style-picker closed, and change the aria-expanded.
+ */
+const closed = () => {
+  document.querySelector('#trigger').setAttribute('aria-expanded', 'false');
 };
 
 /**
@@ -114,10 +126,15 @@ const changeIcon = (ev) => {
 const changePictogram = (ev) => {
   const pictogramName = ev.detail.value;
   const flatPictograms = pictograms.flatMap((group) => group.items);
-  const pictogram = flatPictograms.find((item) => item.value === pictogramName);
+  const pictogram = flatPictograms?.find(
+    (item) => item.value === pictogramName
+  )?.pictogram;
 
   const pictogramHolderEl = document.getElementById('inline-tile-pictogram');
-  const pictogramTemplate = renderCarbonPictogram(pictogram.pictogram);
+  const pictogramTemplate = renderCarbonPictogram({
+    ...pictogram,
+    attrs: { ...pictogram.attrs, 'aria-label': pictogramName },
+  });
   const container = document.createElement('div');
   render(pictogramTemplate, container);
   pictogramHolderEl.innerHTML = '';
@@ -159,6 +176,9 @@ const argTypes = {
   searchInputPlaceholder: {
     control: 'text',
   },
+  searchLabel: {
+    control: 'text',
+  },
 };
 
 export const ColorAndIcon = {
@@ -171,6 +191,7 @@ export const ColorAndIcon = {
     searchCloseButtonLabel: 'Clear search input',
     emptyStateTitle: 'No results found',
     emptyStateSubtitle: 'Try a different search',
+    searchLabel: 'Search',
   },
   argTypes,
   /**
@@ -186,6 +207,7 @@ export const ColorAndIcon = {
       <div class="style-picker-story-container">
         <cds-layer class="toolbar-layer">
           <clabs-style-picker
+            id="style-picker"
             align=${args.align}
             ?open=${args.open}
             heading=${args.heading}
@@ -193,11 +215,16 @@ export const ColorAndIcon = {
             ?enable-search=${args.enableSearch}
             search-close-button-label=${args.searchCloseButtonLabel}
             empty-state-title=${args.emptyStateTitle}
-            empty-state-subtitle=${args.emptyStateSubtitle}>
+            empty-state-subtitle=${args.emptyStateSubtitle}
+            search-label=${args.searchLabel}
+            @clabs-style-picker-close=${closed}>
             <cds-icon-button
+              id="trigger"
               slot="trigger"
               kind=${BUTTON_KIND.GHOST}
-              @click=${toggleButton}>
+              @click=${toggleButton}
+              aria-expanded="${args.open}"
+              aria-controls="style-picker">
               ${ColorPalette16({ slot: 'icon' })}
               <span slot="tooltip-content">Color palette</span>
             </cds-icon-button>
@@ -270,6 +297,7 @@ export const ColorAndPictogram = {
     searchCloseButtonLabel: 'Clear search input',
     emptyStateTitle: 'No results found',
     emptyStateSubtitle: 'Try a different search',
+    searchLabel: 'Search',
   },
   argTypes,
   /**
@@ -285,6 +313,7 @@ export const ColorAndPictogram = {
       <div class="style-picker-story-container">
         <cds-layer class="toolbar-layer">
           <clabs-style-picker
+            id="style-picker"
             align=${args.align}
             ?open=${args.open}
             heading=${args.heading}
@@ -292,11 +321,16 @@ export const ColorAndPictogram = {
             ?enable-search=${args.enableSearch}
             search-close-button-label=${args.searchCloseButtonLabel}
             empty-state-title=${args.emptyStateTitle}
-            empty-state-subtitle=${args.emptyStateSubtitle}>
+            empty-state-subtitle=${args.emptyStateSubtitle}
+            search-label=${args.searchLabel}
+            @clabs-style-picker-close=${closed}>
             <cds-icon-button
+              id="trigger"
               slot="trigger"
               kind=${BUTTON_KIND.GHOST}
-              @click=${toggleButton}>
+              @click=${toggleButton}
+              aria-expanded="${args.open}"
+              aria-controls="style-picker">
               ${ColorPalette16({ slot: 'icon' })}
               <span slot="tooltip-content">Pictogram list</span>
             </cds-icon-button>
@@ -351,7 +385,13 @@ export const ColorAndPictogram = {
         <div class="inline-tile-holder">
           <cds-tile class="inline-tile">
             <div id="inline-tile-pictogram">
-              ${renderCarbonPictogram(pictograms[0].items[0].pictogram)}
+              ${renderCarbonPictogram({
+                ...AmsterdamWindmill,
+                attrs: {
+                  ...AmsterdamWindmill.attrs,
+                  'aria-label': AmsterdamWindmill.name,
+                },
+              })}
             </div>
             <br />
             <h6>Primary text</h6>